Revert "Merge remote-tracking branch 'origin/master' into LargeQueuePublishing"
This reverts commit d8f212fe895ccfb9eac60aa3acf2418fcf4f17e7, reversing
changes made to 409a51b23736fe24446b1f1e6a3e87e705401f11.

+++ Stack/Opc.Ua.Core/Stack/State/NodeState.cs
@@ -3669,7 +3669,7 @@ namespace Opc.Ua
 
                     if (ServiceResult.IsGood(result))
                     {
-                        value = (ushort)accessRestrictions;
+                        value = accessRestrictions;
                     }
 
                     if (value != null || result != null)
